move makeExternal() to src/stateInMemory.js
